agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
openai>=1.50.0
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
jira>=3.8.0
